Refactor collection form to allow for better composition.
