+++ ui/apps/platform/src/Containers/Collections/CollectionsPage.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
+import { PageSection } from '@patternfly/react-core';
 
 import usePermissions from 'hooks/usePermissions';
 import useURLParameter from 'hooks/useURLParameter';
@@ -26,10 +27,14 @@ function CollectionsPage() {
 
     if (validPageActionProp) {
         return (
-            <CollectionsFormPage
-                hasWriteAccessForCollections={hasWriteAccessForCollections}
-                pageAction={validPageActionProp}
-            />
+            <>
+                <PageSection className="pf-u-h-100" padding={{ default: 'noPadding' }}>
+                    <CollectionsFormPage
+                        hasWriteAccessForCollections={hasWriteAccessForCollections}
+                        pageAction={validPageActionProp}
+                    />
+                </PageSection>
+            </>
         );
     }
